Fixed bug in breakStringIntoLines
There was a problem case when the line had been broken and we ended up within the forwardWindow of the end of the quote, or really right at the end of the quote. 
This caused the backward search to happen, which messed up the line break. This quote is a great example of the problem:
"You are who you are, and no one can change that. So be yourself, because no one can do it better." 
Without the fix, the "better" was incorrectly ending up on a line of its own.