Add sleep/wake and backup network support
Sleep is currently set for 1am, wake at 7am. 
Tap when asleep to wake back up. 
This involved adding the time lib to track the time and checking it periodically to see if it is sleep or wake time. (currently about every 10 mins) 

Also brought over the connectWifi function from the round glance display.